Add custom tag support and mark inline tag
